Validation of PLACE command is completed and separate validators for Face and Coordinates are included
